Use smaller transactions Evict caches right after assigning keypair
